[BACKEND] Refactor after add new class
